Refactor place of default editor control settings
